+++ modules/dbfs/databaseMock.go
@@ -287,7 +287,7 @@ func (dm *DatabaseMock) MySQLProjectGetFiles(projectID int64) ([]FileMeta, error
 }
 
 // MySQLProjectGrantPermission is a mock of the real implementation
-func (dm *DatabaseMock) MySQLProjectGrantPermission(projectID int64, grantUsername string, permissionLevel int8, grantedByUsername string) error {
+func (dm *DatabaseMock) MySQLProjectGrantPermission(projectID int64, grantUsername string, permissionLevel int, grantedByUsername string) error {
 	dm.FunctionCallCount++
 	found := false
 
@@ -341,7 +341,7 @@ func (dm *DatabaseMock) MySQLProjectRevokePermission(projectID int64, revokeUser
 }
 
 // MySQLUserProjectPermissionLookup returns the permission level of `username` on the project with the given projectID
-func (dm *DatabaseMock) MySQLUserProjectPermissionLookup(projectID int64, username string) (int8, error) {
+func (dm *DatabaseMock) MySQLUserProjectPermissionLookup(projectID int64, username string) (int, error) {
 	dm.FunctionCallCount++
 	for _, proj := range dm.Projects[username] {
 		if proj.ProjectID == projectID {
